code_agent_implementation/requirements.txt
langchain>=0.0.335
langchain-openai>=0.0.2
pydantic>=2.4.2
python-dotenv>=1.0.0 